chore: renamed tsky to agent
