+++ packages/client/src/agent/index.ts
@@ -1,2 +1,1 @@
-export * from './bsky';
 export * from './agent';
