+++ packages/client/src/bsky/feed/feed.ts
@@ -2,7 +2,7 @@ import type {
   AppBskyFeedGetFeed,
   AppBskyFeedSendInteractions,
 } from '@tsky/lexicons';
-import type { Client } from '~/tsky/client';
+import type { Client } from '~/agent/client';
 import type { RPCOptions } from '~/types';
 import { Paginator } from '~/utils';
 import { FeedGenerator } from './generator';
