+++ packages/client/src/bsky/feed/generator.ts
@@ -3,7 +3,7 @@ import type {
   AppBskyFeedGetFeedGenerators,
   AppBskyFeedGetFeedSkeleton,
 } from '@tsky/lexicons';
-import type { Client } from '~/tsky/client';
+import type { Client } from '~/agent/client';
 import type { RPCOptions } from '~/types';
 import { Paginator } from '~/utils';
