+++ packages/client/src/bsky/post/post.ts
@@ -5,7 +5,7 @@ import type {
   AppBskyFeedGetRepostedBy,
   AppBskyFeedSearchPosts,
 } from '@tsky/lexicons';
-import type { Client } from '~/tsky/client';
+import type { Client } from '~/agent/client';
 import type { RPCOptions } from '~/types';
 import { Paginator } from '~/utils';
