+++ packages/client/src/starterpack/starterpack.ts
@@ -2,7 +2,7 @@ import type {
   AppBskyGraphGetStarterPack,
   AppBskyGraphGetStarterPacks,
 } from '@tsky/lexicons';
-import type { Client } from '~/tsky/client';
+import type { Client } from '~/agent/client';
 import type { RPCOptions } from '~/types';
 import { Paginator } from '~/utils';
