--- packages/client/src/tsky/index.ts
@@ -1,1 +1,0 @@
-export * from './tsky';
