+++ packages/client/src/video/video.ts
@@ -1,5 +1,5 @@
 import type { AppBskyVideoDefs, AppBskyVideoUploadVideo } from '@tsky/lexicons';
-import type { Client } from '~/tsky/client';
+import type { Client } from '~/agent/client';
 import type { RPCOptions } from '~/types';
 
 export class Video {
